docker-build: move build_args logic into separate conditional
Cleans it up, allows easier quoting

Disable shellcheck for optional parameter expansion breaking quoted
variables. The unquoted part will be null if it's not defined, otherwise
it will be a quoted variable.

There might be a way to clean this up. When optional, it should not include
the quotes, but unquoted won’t pass shellcheck